README.md
This is a plugin for CKEditor. It can be found at: http://ckeditor.com/addon/inlinesave

This plugin allows the user to save the content for an inline editor. (The normal save button is disabled for inline mode, for compatibility reasons.) When the save button is clicked, the data in the editor is sent to a php file via $_POST with Jquery Ajax. This allows the developer to collect data from an inline editor much in the same way it can be collected from the regular editor.