Update plugin.js
I think an error for onSave should only be thrown if the return value (sendDataOk) is explicitly set to false.  No return value being set (== "undefined) should not cause an error to be thrown.  Since "undefined" and false are both falsey values, must differentiate between them.  However, is throwing an error always the best approach?  How about checking of onFailure is set and calling that with a xhttp.status semaphore value of -1 if the save was aborted?

Also the else is not needed bfore the CKEDITOR.tools.extend line  because the return exits the function if the if is true.